+++ test/unit/enumerator_builder_test.rb
@@ -78,6 +78,15 @@ module JobIteration
       )
     end
 
+    test_builder_method(:build_parallel_enumerator) do
+      enumerator_builder.build_parallel_enumerator(
+        instances: 2,
+        cursor: { "instance" => 0, "inner_cursor" => nil },
+      ) do |_instance, _instances, inner_cursor|
+        enumerator_builder.build_array_enumerator([1, 2, 3], cursor: inner_cursor)
+      end
+    end
+
     # checks that all the non-alias methods were tested
     raise "methods not tested: #{methods.inspect}" unless methods.empty?